wx-event-reliability/pyproject.toml
[project]
name = "wx_event_reliability"
version = "0.1.0"
description = "Agentic weather app: geocoding, rich date parsing, Open-Meteo fetch, and physics RAG."
authors = [{ name = "Yavar Khan", email = "yavarkhan1997@gmail.com" },{name = "Ishita Srivastava", email = "ishi961992@gmail.com"}]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
# Core ADK + Vertex AI (match academic example)
google-adk = "^1.0.0"
google-genai = "^1.9.0"
google-cloud-aiplatform = { version = "^1.93", extras = ["adk", "agent-engines"] }
python-dotenv = "^1.0.1"
pydantic = "^2.10.6"

# Weather tools
requests = "^2.32.3"
dateparser = "^1.2.0"
timezonefinder = "^6.5.6"

# Optional (if you want to serve an API)
fastapi = "^0.115.0"

# Optional (only if you’re generating PDFs locally)
reportlab = "^4.2.2"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
google-adk = { extras = ["eval"], version = "^1.0.0" }
google-cloud-aiplatform = { version = "^1.93", extras = ["adk", "agent-engines", "evaluation"] }
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
black = "^25.1.0"
pandas = "^2.2.3"
tabulate = "^0.9.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"




